pyproject.toml
[tool.poetry]
name = "hw_29"
version = "0.1.0"
description = ""
authors = ["Harry <vin-ig@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"
drf-spectacular = "^0.23.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
